refactor: apply code-organization conventions to trading service
Codebase-wide audit against the conventions established in #34
(docs/rules/conventions.md). Only the trading service needed changes;
all other in-scope packages were audited clean.

- Split internal/trading/pg_repository.go (323 lines, three resources)
  per the aggregator pattern: pg_repository.go keeps the struct +
  constructor; orders, trades, and balances methods move to
  pg_repository_<resource>.go. statusSlice moves with its only
  consumer (orders). Pure move, no behavior change.
- Extract decryptHMACKey from authenticateAPIKey in
  internal/trading/auth/middleware.go: the decrypt + base64-decode
  chain becomes a pure (value, error) helper, keeping status-code
  decisions at the call site. The timestamp block stays inline because
  its two failure modes differ in WithAuthFailureHook semantics.

Refs #77
